ships: promote ship predicates and queries to member methods on Ship
Promote ship predicates, capacities, and domain queries to first-class
member methods on the `Ship` class with single-line Doxygen comments:
- `is_docked()`: computed from `docked` flag and `LEVEL_SHIP` destination.
- `is_landed()`: computed from `docked` flag and `LEVEL_PLAN` destination.
- `is_laser_on()`: computed from `laser` equipment and `fire_laser`.
- `is_hyper_drive_ready()`: delegated to `HyperDriveData::is_ready()`.
- `is_overloaded()`: validation against resource/fuel/crew/destruct limits.
- `has_switch()`, `can_bombard()`, `can_navigate()`, `can_aim()`, `has_sight()`.
- `effective_armor()`, `active_guns()`, `shipbody()`, `hanger_space()`.
- Capacity queries: `available_crew()`, `available_mil()`,
  `max_crew_capacity()`, `max_resource_capacity()`,
  `max_fuel_capacity()`, `max_destruct_capacity()`, `max_speed_capacity()`,
  `repair_capacity()`, `effective_cost()`.

Positioned the `Shipdata` template table before `class Ship` in
`gblib-ships.cppm` to enable direct inlining of type-specific ability
lookups. Forwarded existing free functions in `ships.cc` to call the new
member methods.

Testing:
- Added comprehensive unit tests in `gblib_test.cc` covering standard
  battleship and factory specialization predicates and capacities.
- Verified 156/156 tests passing (100%).

+++ gb/entities/shootblast.cc
@@ -275,7 +275,7 @@ static std::pair<int, std::string> do_radiation(Ship& ship, double tech,
   double fac = (2. / 3.14159265) *
                std::atan((double)(5 * (tech + 1.0) / (ship.tech() + 1.0)));
 
-  int arm = std::max(0UL, armor(ship) - hits / 5);
+  int arm = std::max(0, static_cast<int>(armor(ship)) - hits / 5);
   int body = shipbody(ship);
 
   int penetrate = 0;
@@ -320,7 +320,7 @@ do_damage(EntityManager& em, player_t who, Ship& ship, double tech,
     }
 
   double fac = p_factor(tech, ship.tech());
-  int arm = std::max(0UL, armor(ship) + defense - hits / 5);
+  int arm = std::max(0, static_cast<int>(armor(ship)) + defense - hits / 5);
   double body = std::sqrt((double)(0.1 * shipbody(ship)));
 
   int critdam = 0;
